Updated some functionality needed for HouseState, also hand coded house state like a lunatic, also committing incorrectly, also I'm tired

+++ api/api/api_build/models.py
@@ -163,7 +163,7 @@ class Powerusage(models.Model):
     timestamp = models.DateTimeField()
     sensorid = models.IntegerField()
     endtimestamp = models.DateTimeField(blank=True, null=True)
-    usage = models.IntegerField()
+    usage = models.FloatField()
     cost = models.FloatField()
 
     class Meta:
